Fix repository rest resource issue

+++ src/main/java/com/funshopper/FunshopperDomain/model/Category.java
@@ -1,17 +1,10 @@
 package com.funshopper.FunshopperDomain.model;
 
-import java.util.Set;
-
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.OneToMany;
-import javax.persistence.Table;
-
 import lombok.Getter;
 import lombok.Setter;
+
+import javax.persistence.*;
+import java.util.Set;
 
 @Getter
 @Setter
@@ -21,7 +14,7 @@ public class Category {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private Integer id;
+	private Long id;
 	
 	private String name;
